requirements.txt
boto3==1.22.8
botocore==1.25.8
fastapi==0.103.0
mangum==0.17.0
numpy==1.24.3
pandas==2.0.3
pydantic==1.10.12
SQLAlchemy==2.0.10
PyMySQL==1.1.0
pyjwt==2.8.0
requests==2.28.1
uvicorn==0.23.2